SIL-80 complete search function
