+++ views/layouts/main.handlebars
@@ -29,11 +29,13 @@
             </svg>
         </div>
         <div class="searchBox">
-            <input type="text" class = "search" placeholder="Search">
-          
-            <svg width="29" height="20" viewBox="0 0 32 36" fill="none" xmlns="http://www.w3.org/2000/svg" id="search-btn">
-                <path fill-rule="evenodd" clip-rule="evenodd" d="M18.3269 23.6537C13.1064 26.3752 6.52489 25.0948 2.74421 20.3032C-1.5328 14.8827 -0.60575 7.0212 4.81483 2.74419C10.2354 -1.53282 18.0969 -0.605772 22.3739 4.81481C26.1546 9.60643 25.8689 16.3053 22.0075 20.7497L30.8557 31.9637C31.6577 32.98 31.4838 34.454 30.4675 35.256C29.4511 36.0579 27.9771 35.8841 27.1752 34.8677L18.3269 23.6537ZM17.8675 19.3193C14.1414 22.2456 8.74768 21.6048 5.81136 17.8834C2.87091 14.1568 3.50826 8.752 7.23491 5.81156C10.9616 2.87111 16.3663 3.50846 19.3068 7.23511C22.2431 10.9565 21.6116 16.3514 17.8989 19.2946C17.8936 19.2986 17.8884 19.3027 17.8831 19.3069C17.8779 19.311 17.8727 19.3151 17.8675 19.3193Z" fill="#ADADAD"/>
-            </svg>
+            <input type="text" class = "search" id="search-input" placeholder="Search">
+            <button id="search-btn">
+                <svg width="29" height="20" viewBox="0 0 32 36" fill="none" xmlns="http://www.w3.org/2000/svg" >
+                    <path fill-rule="evenodd" clip-rule="evenodd" d="M18.3269 23.6537C13.1064 26.3752 6.52489 25.0948 2.74421 20.3032C-1.5328 14.8827 -0.60575 7.0212 4.81483 2.74419C10.2354 -1.53282 18.0969 -0.605772 22.3739 4.81481C26.1546 9.60643 25.8689 16.3053 22.0075 20.7497L30.8557 31.9637C31.6577 32.98 31.4838 34.454 30.4675 35.256C29.4511 36.0579 27.9771 35.8841 27.1752 34.8677L18.3269 23.6537ZM17.8675 19.3193C14.1414 22.2456 8.74768 21.6048 5.81136 17.8834C2.87091 14.1568 3.50826 8.752 7.23491 5.81156C10.9616 2.87111 16.3663 3.50846 19.3068 7.23511C22.2431 10.9565 21.6116 16.3514 17.8989 19.2946C17.8936 19.2986 17.8884 19.3027 17.8831 19.3069C17.8779 19.311 17.8727 19.3151 17.8675 19.3193Z" fill="#ADADAD"/>
+                </svg>
+            </button>
+            
     
             
         </div>
@@ -103,6 +105,7 @@
     <script src="https://cdn.jsdelivr.net/npm/handlebars/dist/handlebars.min.js"></script>
     
     <script src="/js/main.js"></script>
+    
 </body>
 
 </html>
