--- views/searchProduct.handlebars
@@ -1,189 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <!-- Bootstrap CSS -->
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-  <link rel="stylesheet" href="public\assets\homepage.css">
-  <title>Document</title>
-</head>
-<style>
-
-</style>
-
-<body>
-  <nav class="navbar bg-dark navbar-dark">
-    <div class="container-fluid">
-      <div class="row">
-        <div class="col-4 position-relative ">
-          <a href="" class="navbar-brand"><img class="w-75 " src="public\assets\images\icon\logo.svg" alt="Logo"></a>
-        </div>
-        <div class="col-8">
-          <div class="position-absolute start-50  ">
-            <form class="d-flex" role="search">
-              <button class="bg-dark border-0 " type="submit"><img class="bg-dark w-75 " src="public\assets\images\icon-s.svg"
-                  alt=""></button>
-              <input class="form-control me-2 bg-dark p-0 input border-top-0 border-start-0 border-end-0 rounded-0  "
-                type="search" placeholder="Search for product" aria-label="Search">
-            </form>
-          </div>
-        </div>
-      </div>
-
-    </div>
-  </nav>
-
-  <section>
-
-    <div class="row mt-4">
-      <div class="col-md-4">
-        <div>
-          <img class="img-fluid " src="public\assets\images\icon\imageicon.svg" alt="">
-        </div>
-        <div class="mt-4 ms-4 p">
-          <div class="accordion accordion-flush" id="accordionFlushExample">
-            <div class="accordion-item">
-              <h2 class="accordion-header" id="flush-headingOne">
-                <button class="accordion-button buttonf collapsed text-danger fw-bold " type="button"
-                  data-bs-toggle="collapse" data-bs-target="#flush-collapseOne" aria-expanded="false"
-                  aria-controls="flush-collapseOne">
-                  category_name
-                </button>
-              </h2>
-              <div id="flush-collapseOne" class="accordion-collapse collapse" aria-labelledby="flush-headingOne"
-                data-bs-parent="#accordionFlushExample">
-                <div class="accordion-body buttons">
-                  <h6>product_name</h6>
-                  <h6 class="">product_name</h6>
-                </div>
-              </div>
-            </div>
-            <div class="accordion-item">
-              <h2 class="accordion-header" id="flush-headingTwo">
-                <button class="accordion-button collapsed text-danger fw-bold " type="button" data-bs-toggle="collapse"
-                  data-bs-target="#flush-collapseTwo" aria-expanded="false" aria-controls="flush-collapseTwo">
-                  category_name
-                </button>
-              </h2>
-              <div id="flush-collapseTwo" class="accordion-collapse collapse" aria-labelledby="flush-headingTwo"
-                data-bs-parent="#accordionFlushExample">
-                <div class="accordion-body buttons">
-                  <h6>product_name</h6>
-                  <h6 class="">product_name</h6>
-                </div>
-              </div>
-            </div>
-            <div class="accordion-item">
-              <h2 class="accordion-header" id="flush-headingThree">
-                <button class="accordion-button collapsed text-danger fw-bold " type="button" data-bs-toggle="collapse"
-                  data-bs-target="#flush-collapseThree" aria-expanded="false" aria-controls="flush-collapseThree">
-                  category_name
-                </button>
-              </h2>
-              <div id="flush-collapseThree" class="accordion-collapse collapse" aria-labelledby="flush-headingThree"
-                data-bs-parent="#accordionFlushExample">
-                <div class="accordion-body buttons">
-                  <h6>product_name</h6>
-                  <h6 class="">product_name</h6>
-                </div>
-              </div>
-            </div>
-            <div class="accordion-item">
-              <h2 class="accordion-header" id="flush-headingThree">
-                <button class="accordion-button collapsed text-danger fw-bold " type="button" data-bs-toggle="collapse"
-                  data-bs-target="#flush-collapsefour" aria-expanded="false" aria-controls="flush-collapsefour">
-                  category_name
-                </button>
-              </h2>
-              <div id="flush-collapsefour" class="accordion-collapse collapse" aria-labelledby="flush-headingfour"
-                data-bs-parent="#accordionFlushExample">
-                <div class="accordion-body buttons">
-                  <h6>product_name</h6>
-                  <h6 class="">product_name</h6>
-                </div>
-              </div>
-            </div>
-            <div class="accordion-item">
-              <h2 class="accordion-header" id="flush-headingfive">
-                <button class="accordion-button buttonl collapsed text-danger fw-bold " type="button"
-                  data-bs-toggle="collapse" data-bs-target="#flush-collapsefive" aria-expanded="false"
-                  aria-controls="flush-collapsefive">
-                  category_name
-                </button>
-              </h2>
-              <div id="flush-collapsefive" class="accordion-collapse collapse" aria-labelledby="flush-headingfive"
-                data-bs-parent="#accordionFlushExample">
-                <div class="accordion-body buttons">
-                  <h6>product_name</h6>
-                  <h6 class="">product_name</h6>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-
-      </div>
-
-      <div class="col-md-8 ">
-        <div class="">
-          <div class="row mt-4 ms-5 ms-ms-0">
-            <div class="col-md-4 t ">
-
-              <h6>product_name</h6>
-              <div class="bg-secondary box "></div>
-            </div>
-            <div class="col-md-4">
-              <h6>product_name</h6>
-              <div class="bg-secondary box "></div>
-            </div>
-            <div class="col-md-4">
-              <h6>product_name</h6>
-              <div class="bg-secondary box "></div>
-            </div>
-
-          </div>
-          <div class="row mt-4 ms-5 ms-ms-0">
-            <div class="col-md-4">
-              <h6>product_name</h6>
-              <div class="bg-secondary box"></div>
-            </div>
-            <div class="col-md-4">
-              <h6>product_name</h6>
-              <div class="bg-secondary box"></div>
-            </div>
-            <div class="col-md-4">
-              <h6>product_name</h6>
-              <div class="bg-secondary box"></div>
-            </div>
-
-          </div>
-          <div class="row  mt-4 ms-5 ms-ms-0">
-            <div class="col-md-4">
-              <h6>product_name</h6>
-              <div class="bg-secondary box"></div>
-            </div>
-            <div class="col-md-4">
-              <h6>product_name</h6>
-              <div class="bg-secondary box"></div>
-            </div>
-            <div class="col-md-4">
-              <h6>product_name</h6>
-              <div class="bg-secondary box"></div>
-            </div>
-
-          </div>
-        </div>
-      </div>
-    </div>
-  </section>
-
-  <!-- Bootstrap Bundle with Popper -->
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
-    crossorigin="anonymous"></script>
-</body>
-
-</html>
