+++ views/searchResult.handlebars
@@ -17,4 +17,4 @@
 
 </section>
 
-<script src="/js/searchProduct.js"></script>
+<script src="/js/product.js"></script>
